Create TeamsPersonal-Remediation.ps1
If the detection script returns an Exit 1, this Remediation is triggered to run, which removes the appx package and provisioned packaged to ensure the app is removed. 

NOTE: Post future feature upgrades the app is reinstalled, but with a remediation, the app is removed again within the next cycle of the remediation script running.